Moved the login code to the pubs_ui app level and put into a views.py file. So now it is no longer associated with the pubswh blueprint.
