+++ pubs_ui/pubswh/templates/pubswh/base.html
@@ -92,7 +92,7 @@
                 	</ul>
                 </li>
                 <li><a href="{{ url_for('pubswh.contact') }}">Contact</a></li>
-                {% if current_user.is_authenticated() %}<li><a href="{{ url_for('pubswh.logout_page') }}">Logout</a></li>{% endif %}
+                {% if current_user.is_authenticated() %}<li><a href="{{ url_for('logout_page') }}">Logout</a></li>{% endif %}
             </ul>
             <section id="center">
                 <form style="display: inline;" action="{{ url_for('pubswh.search_results') }}" method="get" name="q">
