+++ pubs_ui/pubswh/templates/pubswh/login.html
@@ -35,4 +35,4 @@
 
 {% endblock %}
 
-{% block url %}<a href="{{ url_for('pubswh.login_page') }}">http://pubs.er.usgs.gov{{ url_for('pubswh.login_page') }}</a>{% endblock %}
+{% block url %}<a href="{{ url_for('login_page') }}">http://pubs.er.usgs.gov{{ url_for('login_page') }}</a>{% endblock %}
